added high goal autonomous

+++ src/main/java/frc/team3223/robot2016/RobotConfiguration.java
@@ -317,7 +317,7 @@ public class RobotConfiguration implements ISpeedControllerProvider {
     return new ToggleButton(getRightJoystick(), simpleDriveReverseButton);
   }
 
-  public double getShooterPitch() {
+  public int getShooterPitch() {
     int encoderValue = sensorManager.getShooterRaiserEncoder().get();
     return encoderValue;
   }
